backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
sqlalchemy==2.0.36
pydantic==2.7.4
httpx==0.27.2
python-dotenv==1.0.1
